requirements.txt
streamlit
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-retrievers-bm25
chromadb
PyStemmer
nest_asyncio
python-dotenv